Travis CI - updated roots
