+++ lib/food/food_menu_detail.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
+import 'package:bhavani_shoppe/common/cart_home.dart';
+import 'package:bhavani_shoppe/common/text_style.dart';
 import 'package:bhavani_shoppe/model/foodMenu.dart';
 import 'package:bhavani_shoppe/model/foodMenuItem.dart';
-import 'package:bhavani_shoppe/ui/common/cart_home.dart';
-import 'package:bhavani_shoppe/ui/common/text_style.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
