--- lib/mockData/food_home_screen_data.json
@@ -1,54 +1,0 @@
-[
-  {
-    "name": "Pizza",
-    "count": "5",
-    "icon" : "ic_pizza.png",
-    "color" : "c0ffee",
-    "items":
-    [
-      {
-        "name": "pizza1",
-        "prize": "500"
-      }
-    ]
-  },
-  {
-    "name": "Burger",
-    "count": "10",
-    "icon" : "ic_pizza.png",
-    "color" : "c0ffee",
-    "items":
-    [
-      {
-        "name": "Burger1",
-        "prize": "500"
-      }
-    ]
-  },
-  {
-    "name": "Chinese",
-    "count": "5",
-    "icon" : "ic_pizza.png",
-    "color" : "c0ffee",
-    "items":
-    [
-      {
-        "name": "chinese1",
-        "prize": "500"
-      }
-    ]
-  },
-  {
-    "name": "Mexican",
-    "count": "10",
-    "icon" : "ic_pizza.png",
-    "color" : "c0ffee",
-    "items":
-    [
-      {
-        "name": "Mexican1",
-        "prize": "500"
-      }
-    ]
-  }
-]
