--- lib/model/cartItem.dart
@@ -1,9 +1,0 @@
-class CartItem {
-  String id;
-  String foodType;
-  String name;
-  int price;
-  int count;
-
-  CartItem({this.id, this.foodType, this.name, this.price, this.count});
-}
