--- lib/ui/food/common/separator.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Separator extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return new Container(
-        margin: new EdgeInsets.symmetric(vertical: 8.0),
-        height: 2.0,
-        width: 20.0,
-        color: new Color(0xff00c6ff)
-    );
-  }
-}
